requirements.txt
Flask==2.1.1
Flask-SQLAlchemy==2.5.1
PyJWT==2.3.0
pymysql==1.0.2
Faker==9.7.0
Werkzeug==2.0.2
